+++ assets/views/partials/nav.html
@@ -23,8 +23,8 @@
                 Artists
             </a>
 
-            <a class="navbar-item" href="/search">
-                Search
+            <a class="navbar-item" href="/artworks">
+                Artworks
             </a>
 
             <a class="navbar-item" href="/guestbook" hx-get="/guestbook" hx-target="#mc-area">
